requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
google-api-core==2.24.0
google-auth==2.36.0
google-cloud-language==2.15.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
idna==3.10
numpy==2.2.0
opencv-python-headless==4.10.0.84
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
